customer-support/requirements.txt
streamlit==1.38.0
psycopg2-binary
qdrant-client
llama-index
rank-bm25
pandas
numpy
sentence-transformers
tqdm
python-dotenv
openai